# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Express Server API",
    "description": "Small Express server API. Multi-schema: pass schema per request where applicable.",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/", "description": "Current server" }],
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "description": "Returns a simple JSON response indicating the server is running",
        "operationId": "getRoot",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Express server is running" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Returns 200 OK if the server is healthy",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": { "type": "string", "example": "OK" }
              }
            }
          }
        }
      }
    },
    "/db-status": {
      "get": {
        "summary": "Database status",
        "description": "Checks MySQL RDS connection pool. Pass schema to ping a specific database/schema; omit for server-level ping.",
        "operationId": "getDbStatus",
        "parameters": [
          {
            "name": "schema",
            "in": "query",
            "required": false,
            "description": "Database/schema name to ping (e.g. tenant_a). Omit to check server connectivity only.",
            "schema": { "type": "string", "nullable": true }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "schema": { "type": "string", "nullable": true, "description": "Schema that was pinged, or null" },
                    "ok": { "type": "boolean", "description": "Whether the DB/schema is reachable" },
                    "latencyMs": { "type": "number", "description": "Round-trip time in milliseconds" },
                    "error": { "type": "string", "description": "Error message if ok is false" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/schema-dump": {
      "post": {
        "summary": "Create full Liquibase changelog for a schema and upload to S3",
        "description": "Generates a full Liquibase changelog (generate-changelog) for the given schema, uploads the file to S3, and deletes the local file.",
        "operationId": "createSchemaDump",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "schema": { "type": "string", "description": "Schema/database name to dump" }
                },
                "required": ["schema"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dump created and uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "schema": { "type": "string" },
                    "s3Key": { "type": "string", "description": "S3 object key where the changelog was stored" },
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/schema-restore": {
      "post": {
        "summary": "Create a new schema and apply a Liquibase changelog from S3",
        "description": "Creates the given schema/database if it does not exist, downloads the changelog file from S3, runs Liquibase update against that schema, and returns basic status.",
        "operationId": "restoreSchemaFromDump",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "schema": { "type": "string", "description": "New schema/database name to create and restore into" },
                  "s3Key": { "type": "string", "description": "S3 object key of the Liquibase changelog file" }
                },
                "required": ["schema", "s3Key"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Schema created/restored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "schema": { "type": "string" },
                    "s3Key": { "type": "string" },
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}